Adds morkgetc(), comments, variable name changes.
Changes all input from fgetc() to morkgetc(). At this point
morkgetc() just calls fgetc() but in the future this could
be used to implement groups since the body of the group
could be buffered, then if not ~abort`'d it could then be
parsed by having morkgetc() read the buffer instead of
calling fgetc() until the buffer was empty.

Changes some names of 'oid' variables to 'object id' names
to make it a little clearer what the are.