Makes group parsing controllable at runtime.
Adds the global variable 'morkDoNotParseGroups' that if set
true will cause the parser to treat any '@' surrounded strings
as meta strings to be ignored.

This will result in all groups being loaded even if aborted.

It will also result in faster operation because the entire
content of the group will not need to be 'ungetted' and the
'regetted'.