Adds command line option '-g' to block groups
Putting the '-g' command line option on will result in the
group header and footer strings being ignored as meta strings.

This will cause all groups to be loaded, even those that have
been aborted in their footers.

It will, however, result in faster runtime because the contents
of the group will not need to be "ungetc"ed and then read a second
time. It might also prevent failure if the support library or
memory available limits the number of characters that ungetc()
can buffer.